feat: 🎸 add overwrite flag to expand and shrink command

+++ packages/sfpowerscripts-cli/src/commands/sfpowerscripts/dependency/expand.ts
@@ -29,6 +29,11 @@ export default class Expand extends SfpowerscriptsCommand {
             description: messages.getMessage('devhubAliasFlagDescription'),
             default: 'HubOrg',
         }),
+        overwrite: flags.boolean({
+            char: 'o',
+            description: messages.getMessage('overWriteProjectConfigFlagDescription'),
+            default: false,
+        }),
         loglevel: flags.enum({
             description: 'logging level for this command invocation',
             default: 'info',
@@ -53,6 +58,7 @@ export default class Expand extends SfpowerscriptsCommand {
     public async execute() {
         let sfpOrg: SFPOrg;
         let defaultProjectConfigPath = './project-config';
+        let projectConfigFilePath: string;
         if (this.flags.devhubalias) sfpOrg = await SFPOrg.create({ aliasOrUsername: this.flags.devhubalias });
         try {
             //Validate dependencies in sfdx-project.json // Read Manifest
@@ -64,13 +70,19 @@ export default class Expand extends SfpowerscriptsCommand {
             projectConfig = await transitiveDependencyResolver.resolveDependencies(Stage.EXPAND);
 
             //Clean up temp directory
-            if (fs.existsSync(defaultProjectConfigPath)) rimraf.sync(defaultProjectConfigPath);
+            if (!fs.existsSync(defaultProjectConfigPath)) fs.mkdirpSync(defaultProjectConfigPath);
 
-            fs.mkdirpSync(defaultProjectConfigPath);
-            let projectConfigFilePath: string = path.join(defaultProjectConfigPath, `sfdx-project.exp.json`);
-            fs.writeFileSync(projectConfigFilePath, JSON.stringify(projectConfig, null, 4));
+            if(this.flags.overwrite){
+                console.log(`Overwriting sfdx-project.json with expanded project config file`);
+                projectConfigFilePath = `sfdx-project.json`;
+                fs.writeFileSync(projectConfigFilePath, JSON.stringify(projectConfig, null, 4));
+                console.log('sfdx-project.json has been updated.')
+            }else{
+                projectConfigFilePath = path.join(defaultProjectConfigPath, `sfdx-project.exp.json`);
+                fs.writeFileSync(projectConfigFilePath, JSON.stringify(projectConfig, null, 4));
+                console.log(`Generated project config file has been saved to ${projectConfigFilePath}`);
+            }
 
-            console.log(`Generated project config file has been saved to ${projectConfigFilePath}`);
         } catch (error) {
             throw new Error('Unable to generate project config file:' + error);
         }
